Extract routes into their own modules
