+++ sockshop/front-end/api/catalogue/index.js
@@ -1,0 +1,24 @@
+(function (){
+  'use strict';
+
+  var express   = require("express")
+    , request   = require("request")
+    , endpoints = require("../endpoints")
+    , helpers   = require("../../helpers")
+    , app       = express()
+
+  app.get("/catalogue/images*", function (req, res, next) {
+    var url = endpoints.catalogueUrl + req.url.toString();
+    request.get(url).pipe(res);
+  });
+
+  app.get("/catalogue*", function (req, res, next) {
+    helpers.simpleHttpRequest(endpoints.catalogueUrl + req.url.toString(), res, next);
+  });
+
+  app.get("/tags", function(req, res, next) {
+    helpers.simpleHttpRequest(endpoints.tagsUrl, res, next);
+  });
+
+  module.exports = app;
+}());
