+++ sockshop/front-end/api/endpoints.js
@@ -1,0 +1,15 @@
+(function (){
+  'use strict';
+
+  module.exports = {
+    catalogueUrl:  "http://catalogue",
+    tagsUrl:       "http://catalogue/tags",
+    cartsUrl:      "http://cart/carts",
+    ordersUrl:     "http://orders",
+    customersUrl:  "http://accounts/customers",
+    addressUrl:    "http://accounts/addresses",
+    cardsUrl:      "http://accounts/cards",
+    loginUrl:      "http://login/login",
+    registerUrl:   "http://login/register"
+  };
+}());
